PubSub-Function-ES/src/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-logging
Elasticsearch
datetime